requirements.txt
asgiref==3.9.2
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coverage==7.10.7
croniter==6.0.0
Django==5.2.6
django-cors-headers==4.9.0
django-redis==6.0.0
django-rq==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
filelock==3.19.1
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.5
requests-file==2.1.0
rpds-py==0.27.1
rq==2.6.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tldextract==5.3.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
zope.event==6.0
zope.interface==8.0.1
